+++ src/core/game.ts
@@ -11,10 +11,8 @@ namespace Lightning {
         private _activateState:State = null;
         private _tweens = new Tween.TweenManager(this);
         private _signals:Signals.SignalManager = new Signals.SignalManager(this);
-        private _physics:any;
-        private _physicsActive:boolean = false;
-        private _physicsWorld:Box2D.Dynamics.b2World;
-        private _physicsWorldBounds:Box2D.Dynamics.b2BodyDef;
+        private _stateManager:StateManager;
+        private _physicsManager:PhysicsManager
         
         // game engine constructor
         constructor(width, height, canvasId:string = 'app') {
@@ -39,6 +37,12 @@ namespace Lightning {
             let renderer = PIXI.autoDetectRenderer(width * scale, height * scale, canvas);
             canvas.style.width = width + 'px';
             canvas.style.height = height + 'px';
+
+            // create the physicsManager 
+            this._physicsManager = new PhysicsManager(this);
+
+            // create the state StateManager
+            this._stateManager = new StateManager(this);
             
             // init the ticker
             this._ticker = PIXI.ticker.shared;
@@ -48,63 +52,10 @@ namespace Lightning {
 
         // gets called on update
         update(time):void {
-            if(this._physicsActive) {
-                this._physicsWorld.Step(1 / 60,  1, 1);
-                this._physicsWorld.ClearForces();
-            }
-            if(this._activateState) {
-                this._activateState.update();
-            }
+            this._physicsManager.update();
             this._tweens.update();
+            this._stateManager.update();
             this._renderer.render(this._world);
-        }
-
-        startState(state, ...params) {
-            let nState = new state(this);
-            this.initState(nState, params);
-        }
-
-        initState(state:State, params) {
-            if(this._activateState === null) {
-                this._world.addChild(state);
-            } else {
-                this._world.removeChild(this._activateState);
-                this._world.addChild(state);
-            }
-            this._activateState = state;
-            state.init(params);
-        }
-
-        startPhysics() {
-            this._physicsWorld = new Box2D.Dynamics.b2World(new Box2D.Common.Math.b2Vec2(0, 10),  true);
-            this._physicsActive = true;
-        }
-
-        collideOnWorldBounds():void {
-            this._physicsWorldBounds = new Box2D.Dynamics.b2BodyDef();
-            let polyFixture:Box2D.Dynamics.b2FixtureDef = new Box2D.Dynamics.b2FixtureDef();
-            polyFixture.shape = new Box2D.Collision.Shapes.b2PolygonShape();
-            polyFixture.density = 1;
-
-            this._physicsWorldBounds = new Box2D.Dynamics.b2BodyDef();
-            this._physicsWorldBounds.type = Box2D.Dynamics.b2Body.b2_staticBody;
-        
-            //down
-            polyFixture.shape.SetAsBox(10, 1);
-            this._physicsWorldBounds.position.Set(9, this.height / 100 + 1);
-            this.physics.CreateBody(this._physicsWorldBounds).CreateFixture(polyFixture);
-            
-            //left
-            polyFixture.shape.SetAsBox(1, 100);
-            this._physicsWorldBounds.position.Set(-1, 0);
-            this.physics.CreateBody(this._physicsWorldBounds).CreateFixture(polyFixture);
-            
-            //right
-            this._physicsWorldBounds.position.Set(this.height / 100, 0);
-            this.physics.CreateBody(this._physicsWorldBounds).CreateFixture(polyFixture);
-            this._physicsWorldBounds.type = Box2D.Dynamics.b2Body.b2_dynamicBody;
-
-            let body = Box2D.Dynamics.b2Body;
         }
 
         generateTexture(...params):any {
@@ -147,10 +98,6 @@ namespace Lightning {
             this._renderer.backgroundColor = val;
         }
 
-        public set state(val:State) {
-            this._activateState = val;
-        }
-
         public get world():PIXI.Container {
             return this._world;
         }
@@ -179,14 +126,13 @@ namespace Lightning {
             return this._signals;
         }
 
-        public get physics():Box2D.Dynamics.b2World {
-            return this._physicsWorld;
+        public get states():StateManager {
+            return this._stateManager;
         }
 
-        public get physicsWorldBounds():Box2D.Dynamics.b2BodyDef {
-            return this._physicsWorldBounds
-        }
-
+        /**
+         * think about refactoring this
+         */
         public get hud():HUD {
             if(!this._hud) {
                 this._hud = new HUD(this);
@@ -195,3 +141,10 @@ namespace Lightning {
         }
     }
 }
+
+/**
+ * Start Ticker
+ * Pause Ticker
+ * Ticker FPS?
+ * Individual State FPS?
+ */