+++ src/core/gameHelper.ts
@@ -1,0 +1,8 @@
+/// <reference path="./../reference.d.ts" />
+
+namespace Lightning {
+
+    export class EngineHelper {
+        
+    }
+}