+++ src/interfaces/iPoint.d.ts
@@ -1,4 +1,0 @@
-declare interface iPoint {
-    x:number;
-    y:number;
-}