+++ src/interfaces/iPointRange.d.ts
@@ -1,6 +1,0 @@
-interface iPointRange {
-    xFrom:number;
-    xTo:number;
-    yFrom:number;
-    yTo:number;
-}