+++ src/interfaces/iPointRange.ts
@@ -1,0 +1,10 @@
+/// <reference path="./../reference.d.ts" />
+
+namespace Lightning {
+    export interface iPointRange {
+        xFrom:number;
+        xTo:number;
+        yFrom:number;
+        yTo:number;
+    }
+}