+++ src/interfaces/iRange.d.ts
@@ -1,4 +1,0 @@
-declare interface iRange {
-    from:number;
-    to:number;
-}