+++ src/reference.d.ts
@@ -1,14 +1,35 @@
+/**
+ * External Libs
+ */
 /// <reference path="../libs/typings/box2d/index.d.ts" />
 /// <reference path="../libs/typings/pixi.js/index.d.ts" />
 /// <reference path="../libs/typings/sound.js/index.d.ts" />
-/// <reference path ="./interfaces/iPoint.d.ts" />
-/// <reference path ="./interfaces/iRange.d.ts" />
-/// <reference path ="./interfaces/iPointRange.d.ts" />
-/// <reference path="./core/state.ts" />
-/// <reference path="./core/game.ts" />
-/// <reference path="./controllers/input.ts" />
-/// <reference path="./utils/font.ts" />
+/**
+ * Interfaces
+ */
+/// <reference path ="./interfaces/iPoint.ts" />
+/// <reference path ="./interfaces/iRange.ts" />
+/// <reference path ="./interfaces/iPointRange.ts" />
+/// <reference path ="./interfaces/iStateMap.ts" />
+/**
+ * Utils
+ */
 /// <reference path="./utils/maths.ts" />
+/// <reference path="./utils/depreciated.ts" />
+/**
+ * Core
+ */
+/// <reference path="./core/game.ts" />
+/// <reference path="./core/states/state.ts" />
+/// <reference path="./core/states/stateManager.ts" />
+/// <reference path="./core/physics/physicsManager.ts" />
+/**
+ * Controllers
+ */
+/// <reference path="./controllers/input.ts" />
+/**
+ * Display
+ */
 /// <reference path="./display/displayObject.ts" />
 /// <reference path="./display/texture.ts" />
 /// <reference path="./display/graphics.ts" />
@@ -16,6 +37,7 @@
 /// <reference path="./display/sprite.ts" />
 /// <reference path="./display/group.ts" />
 /// <reference path="./display/HUD.ts" />
+/// <reference path="./display/bitmapText.ts" />
 /// <reference path="./display/geometry.ts" />
 /// <reference path="./display/hitarea.ts" />
 /// <reference path="./display/button.ts" />